+++ tracpro/msgs/tasks.py
@@ -7,6 +7,7 @@ from django.utils import timezone
 from celery.utils.log import get_task_logger
 from djcelery_transactions import task
 
+from tracpro.client import get_client
 from tracpro.orgs_ext.tasks import OrgTask
 
 
@@ -18,30 +19,34 @@ def send_message(message_id):
     from .models import Message, STATUS_SENT, STATUS_FAILED
 
     message = Message.objects.select_related('org').get(pk=message_id)
+    contacts = [c.uuid for c in message.recipients.all()]
 
-    client = message.org.get_temba_client()
+    client = get_client(message.org)
 
-    try:
-        client.create_broadcast(message.text, contacts=[c.uuid for c in message.recipients.all()])
+    # Can only send up to 100 messages at a time
+    while contacts:
+        try:
+            client.create_broadcast(message.text, contacts=contacts[:100])
+        except Exception:
+            message.status = STATUS_FAILED
+            message.save(update_fields=('status',))
 
-        message.status = STATUS_SENT
-        message.save(update_fields=('status',))
+            logger.error("Sending message %d failed" % message.pk, exc_info=1)
+            return
+        contacts = contacts[100:]
 
-        logger.info("Sent message %d from user #%d" % (message.pk, message.sent_by.pk))
-    except Exception:
-        message.status = STATUS_FAILED
-        message.save(update_fields=('status',))
-
-        logger.error("Sending message %d failed" % message.pk, exc_info=1)
+    logger.info("Sent message %d from user #%d" % (message.pk, message.sent_by.pk))
+    message.status = STATUS_SENT
+    message.save(update_fields=('status',))
 
 
 @task
 def send_unsolicited_message(org, text, contact):
 
-    client = org.get_temba_client()
+    client = get_client(org)
 
     try:
-        client.create_broadcast(text, contacts=[contact.uuid])
+        client.create_broadcast(text=text, contacts=[contact.uuid])
 
         logger.info("Sent unsolicited message response to %s" % (contact.name))
     except Exception:
@@ -55,40 +60,40 @@ class FetchOrgInboxMessages(OrgTask):
         from .models import InboxMessage
         from tracpro.contacts.models import Contact
 
-        client = org.get_temba_client()
+        def update_or_create_message(message, org):
+            contact = Contact.objects.filter(uuid=message.contact.uuid).first()
+            # If the contact sync task hasn't gotten this contact yet,
+            # don't get the message yet
+            if contact:
+                InboxMessage.objects.update_or_create(
+                    rapidpro_message_id=message.id,
+                    org=org,
+                    contact=contact,
+                    defaults={
+                        'text': message.text,
+                        'archived': False,
+                        'created_on': message.created_on,
+                        'sent_on': message.sent_on,
+                        'direction': message.direction[0]
+                        },
+                )
+        client = get_client(org)
 
         # Get non-archived, incoming inbox messages from the past week only
         # because getting all messages was taking too long
         last_week = timezone.now() - relativedelta(days=7)
-        inbox_messages = client.get_messages(_types="I", archived="N", after=last_week)
+        # When this is called by the form, we also want to get the very recent sent messages
+        one_minute_ago = timezone.now() - relativedelta(minutes=1)
+        inbox_messages = client.get_messages(folder='inbox', after=last_week)
+        sent_messages = client.get_messages(folder='sent', after=one_minute_ago)
 
         for inbox_message in inbox_messages:
-            contact = Contact.objects.filter(uuid=inbox_message.contact).first()
-            # If the contact sync task hasn't gotten this contact yet,
-            # don't get the message yet
-            if contact:
-                inbox_message_record = InboxMessage.objects.filter(
-                    rapidpro_message_id=inbox_message.id)
-                if inbox_message_record:
-                    inbox_message_record.update(
-                        org=org,
-                        contact=contact,
-                        text=inbox_message.text,
-                        archived=inbox_message.archived,
-                        created_on=inbox_message.created_on,
-                        delivered_on=inbox_message.delivered_on,
-                        sent_on=inbox_message.sent_on,
-                        direction=inbox_message.direction,
-                    )
-                else:
-                    InboxMessage.objects.create(
-                        org=org,
-                        rapidpro_message_id=inbox_message.id,
-                        contact=contact,
-                        text=inbox_message.text,
-                        archived=inbox_message.archived,
-                        created_on=inbox_message.created_on,
-                        delivered_on=inbox_message.delivered_on,
-                        sent_on=inbox_message.sent_on,
-                        direction=inbox_message.direction,
-                    )
+            update_or_create_message(
+                message=inbox_message,
+                org=org,
+                )
+        for sent_message in sent_messages:
+            update_or_create_message(
+                message=sent_message,
+                org=org,
+                )