+++ tracpro/polls/tasks.py
@@ -1,7 +1,5 @@
 from __future__ import absolute_import, unicode_literals
 
-from collections import OrderedDict
-
 from django.apps import apps
 from django.utils import timezone
 
@@ -13,6 +11,7 @@ from temba_client.utils import parse_iso8601, format_iso8601
 
 from dash.utils import datetime_to_ms
 
+from tracpro.client import get_client
 from tracpro.contacts.models import Contact
 from tracpro.orgs_ext.tasks import OrgTask
 
@@ -31,7 +30,7 @@ class FetchOrgRuns(OrgTask):
         from tracpro.orgs_ext.constants import TaskType
         from tracpro.polls.models import Poll, PollRun, Response
 
-        client = org.get_temba_client()
+        client = get_client(org)
         redis_connection = get_redis_connection()
         last_time_key = LAST_FETCHED_RUN_TIME_KEY % org.pk
         last_time = redis_connection.get(last_time_key)
@@ -48,8 +47,8 @@ class FetchOrgRuns(OrgTask):
 
         total_runs = 0
         for poll in Poll.objects.active().by_org(org):
-            poll_runs = client.get_runs(flows=[poll.flow_uuid], after=last_time, before=until)
-            total_runs += len(poll_runs)
+            poll_runs = client.get_runs(flow=poll.flow_uuid, after=last_time, before=until)
+            total_runs += len(poll_runs.all())
 
             # convert flow runs into poll responses
             for run in poll_runs:
@@ -81,7 +80,7 @@ def pollrun_start(pollrun_id):
         raise ValueError("Can't start non-regional poll")
 
     org = pollrun.poll.org
-    client = org.get_temba_client()
+    client = get_client(org)
 
     contacts = Contact.objects.active()
     if pollrun.pollrun_type == PollRun.TYPE_PROPAGATED:
@@ -91,7 +90,9 @@ def pollrun_start(pollrun_id):
         contacts = contacts.filter(region=pollrun.region)
     contact_uuids = list(contacts.values_list('uuid', flat=True))
 
-    runs = client.create_runs(pollrun.poll.flow_uuid, contact_uuids, restart_participants=True)
+    runs = client.create_flow_start(
+        flow=pollrun.poll.flow_uuid, urns=None, contacts=contact_uuids, restart_participants=True)
+
     for run in runs:
         Response.create_empty(org, pollrun, run)
 
@@ -114,9 +115,10 @@ def pollrun_restart_participants(pollrun_id, contact_uuids):
         raise ValueError("Can only restart last pollrun of poll for a region")
 
     org = pollrun.poll.org
-    client = org.get_temba_client()
+    client = get_client(org)
 
-    runs = client.create_runs(pollrun.poll.flow_uuid, contact_uuids, restart_participants=True)
+    runs = client.create_flow_start(
+        flow=pollrun.poll.flow_uuid, contacts=contact_uuids, restart_participants=True)
     for run in runs:
         Response.create_empty(org, pollrun, run)
 
@@ -140,26 +142,28 @@ def sync_questions_categories(org, polls):
 
     # Save the associated Questions for this poll here
     # now that these polls have been activated for the Org
-    selected_poll_names = [poll.name for poll in polls]
+    flow_uuids = [poll.flow_uuid for poll in polls]
+    total_polls = len(flow_uuids)
 
-    temba_polls = org.get_temba_client().get_flows(archived=False)
-    temba_polls = {p.uuid: p for p in temba_polls}
-
-    total_polls = len(polls)
     logger.info(
         "Retrieving Questions and Categories for %d Poll(s) that were recently updated via the interface." %
         (total_polls))
-    for temba_poll in temba_polls.values():
-        if temba_poll.name in selected_poll_names:
-            poll = polls[selected_poll_names.index(temba_poll.name)]
-            # Sync related Questions, and maintain question order.
-            temba_questions = OrderedDict((r.uuid, r) for r in temba_poll.rulesets)
 
-            # Remove Questions that are no longer on RapidPro.
-            poll.questions.exclude(ruleset_uuid__in=temba_questions.keys()).delete()
+    result = get_client(org).get_definitions(flows=flow_uuids)
+    flows = result.flows
 
-            # Create new or update existing Questions to match RapidPro data.
-            for order, temba_question in enumerate(temba_questions.values(), 1):
-                Question.objects.from_temba(poll, temba_question, order)
+    for flow in flows:
+        poll = polls[flow_uuids.index(flow['metadata']['uuid'])]
+
+        # Sync related Questions, and maintain question order.
+        rule_sets = flow['rule_sets']
+
+        # Remove Questions that are no longer on RapidPro.
+        rule_uuids = [rset['uuid'] for rset in rule_sets]
+        poll.questions.exclude(ruleset_uuid__in=rule_uuids).delete()
+
+        # Create new or update existing Questions to match RapidPro data.
+        for order, rule_set in enumerate(rule_sets, 1):
+            Question.objects.from_temba(poll=poll, temba_question=rule_set, order=order)
 
     logger.info("Completed retrieving Questions and Categories for %d Poll(s)." % (total_polls))